Adding Domain and Entity modules
By splitting our code into multiple modules we can enforce the inward dependency rule, as well as keep our code loosely coupled.

The entities modules will contain simple data classes representing the entities of the application. Because this is a front end application these will be object representations of the data passed through the JSON API. Although not strictly entities as they are immutable data classes, in the context of the mobile applications they are our reference to the application entities. Because they will consists of immutable data objects they will not need any unit test.

The domain module will contain our domain specific code. It should aim to encapsulate the business rules of the mobile application. The most important part of this module is that it has no Android dependencies. This has a few advantages. Firstly it allows us to run our tests in the JVM making them faster. Secondly decoupling our business rules from the application platform allows us to potentially deploy our code to a different platform. Another advantage of decoupling business rules from the platform is that it allows - through dependency inversion - to easily swap out platform specific components such as data storage, networking, location etc when and where we like - for example when a new OS version or library becomes available