Adding URLString Type alias
Our API returns an image for each shift as a url. Rather than typing these properties as Strings we can create type aliases to make the code more readable. For example seeing a property image:String is not very informative but image:URLString tells us semantically what the data is. If the developer wants more information about URLString they can navigate to the definition and see it is merely a string. However going forward when they see this alias they will know both the type and the format