Adding Start/EndShift Requests
As a naming convention I like to use Request for post body data classes. I then use Entity for the response body classes. This avoids confusion about which direction the data is travelling. While I used the ISO8601 type alias for the time property I chose not to do so for the lat/long properties as I felt that name was enough to tell the format of the string as lat/long are a common format.