Adding ShiftEntity
Simple data class that represents a shift. One thing to note is that end, and endLatitude/endLongitude can be empty if the user has not finished their shift. Ideally these properties would be null or the data structured into sub objects. Unfortunately this detail leaves us with no way to express this through the type system. We could potentially create out own subclass of string if this became a common pattern but for this simple exercise it is enough, I think, to simply note it in the commit history