Adding Retrofit dependency
I decided to use retrofit for the networking layer. The reason i chose to use retrofit is that, firstly we always need a library to handle networking. You could opt to roll your own but this is a solved issue and will cause more issues than it solves.

The reason I like retrofit is that it allows you to define your API using an interface. This is important for the architecture. The domain layer does not need to know how the data is retrieved. All it needs in an API to do so. This allows use one of the most powerful software design patterns - Dependency inversion. The domain layer simply provides the API and the client can provide the implementation best suited to it.

One example where this can be extremely useful is when writing UI tests. In this scenario rather than relying on hitting a real server, we can provide an API that goes to a stub sever, or replace the whole API with a compiled hard stubbed version allowing us to run the whole test suite without any external dependencies.

One issue with using retrofit is that although the domain provides the API, it needs to annotate it with retrofit annotations in order for the compiler to generate the concrete implementations. This breaks our dependency rule, by the domain module needing to have a dependency to the retro fit library (ideally it would only declare the interface). However, because retrofit is using annotations this does not stop us providing other implementations. If you were to be purist about it (which you may want to be if developing an SDK for example) you could simply declare the interface and allow the client to build it. One way I have done this previously is to create another interface in the client module (app) and have it use the retrofit annotations. We can then use a delegate pattern to return an anonymous instance of the domains interface and forward calls to our retrofit instance. However I think this is generally overkill for most cases (especially this one) and we can justify the breaking of our dependency rule to allow less boilerplate to be written knowing we can still provide different implementations of the interface, if and when we want