Adding Coroutines dependency
I chose to use kotlin coroutines to handle async operations for 3 reasons. Firstly because I could not use RX java which is my go to library for async operations and reactive programming. Secondly because it is a "natively" provided tool that integrates with retrofit allowing us to compose multiple async operations. Finally I used them because I wanted to learn more about them, having only read about them. Whenever I have to do a challenge like this I always try to learn something new to guarantee that I get something out of the exercise. Coroutines have been around for a while now and it is about time I used them. Mainly I have avoided them because I have been programming in Reactive pattern, and as powerful as they are not as powerful as RX if you are writing reactive code.