Adding ShiftsCache
We will need a method to store shifts once they are loaded. This will allow us to implement cache back up, error handling etc. We can use the suspend keyword to to delegate the async nature of storingthe data to the client as almost always storing data (shared prefs, SQL etc ) will be an async operation