Adding ShiftsMemoryCache
Simple implementation of the ShiftsCache interface. This interface simply stores the data in a local property. While we could easily have implemented this in the client, putting the extremely simple operation behind our interface gives us easy leverage to change how we store the data when requirements change at a future date