Adding SimpleEventEmitter
An initial simple implementation of EventEmitter. This is a basic event listener pattern. We store the listener references as WeakReferences to avoid misuse of the api causing memory leaks (not calling removeListener on tearDown). This does make getting full unit test coverage difficult, ie we cannot test the scenario when a weak reference fails. We could get around this by providing a weakReference factory that we can control from our test but this seems like overkill at this point. If it becomes an issue of causes any bugs we could move to change the constructor to follow this pattern