Adding TestHelper mockFunction method
When dealing with listeners we have to mock functions. In mockk this means we need to explicitly handle the return from the function a call. using relaxedUnitFun does not work. Rather than duplicate the boiler plate we can use this function to do this. because of the ever / any capture we need to know the type specified by T at runtime, hence the need for inline and reified