Adding LatestEventEmitter
Some implementations of the EventEmitter way want listeners to be provided with the latest value the add the listener, rather than only getting subsequent events. This is equivalent to a get and listen. Ti handle this we simply store the last value an immediately send it to listeners when they are added