Adding Jacoco coverage for domain
Adding jacoco coverage task. This is useful for checking and also enforcing code coverage metrics. Sadly because Jacoc covers byte code not source code we have to have quite low metrics. This is because kotlin coroutines generates a lot of byte code that is not coverage in our tests. While jacoco does have filters for compiler generated code it does not (yet) have filters for coroutine generated code. However, I decided to keep the coverage task as it is useful in development, firstly because the lines metric is still able to be 100% but also because despite these facts we should be able to at least maintain a relatively stead statistic. The coverage stats are all relative after all.

I did deliberately exclude the Event class as this is a simple data class that a unit test would not really add any value to.