Refactoring ShiftRepository to have returned deferred wait for load to complete
When calling start/end shift we ideally do not want the Deferred object to return until the load has finished. We can achieve this by having the load method return a deferred object and calling await. This is simply a case of moving the previous contents of loadShifts into its own method that is called from loadShifts, then changing the start/end methods to call this new doLoadAsync method and await the returned Job