Adding Presentation module
This module is intended to house the presentation / framework integration layer of the application. We could provide this as part of the app module, but having a seperate module can be beneficial. One of the problems we face is that when using a dependency injector at some point something needs to know about everything, in order to build the final dependency graph. This means the app module, which is always the top of your gradle dependency graph needs to know about all other modules. This can lead to code in this module knowing about things it should not, which can lead to develops using classes that were not intended to be exposed.

By having a seperate presentation module we can expose only the domain layer, meaning developers can not use entities in the presentation layer by mistakenly. ultimately this is to enforce our dependency rules, but can also have the side effect of incremental compilation gains.

To this end the ShiftsActivity was moved to this package, all that was needed was to update the package name in the app modules AndroidManifest and everything worked as expected