Adding Start Shift Command
This command returns a Deferred object to manage the async nature of its operations, asynchronously getting the users location, combining this with the current time parsed into ISO8601 format before passing to the ShiftsRepository.

Because we support older versions of Android we have a problem formatting ISO8601 dates, in that there is no defined format that will work across OS versions. For older versions the closest we can get is to generate the date without the : in the timezone portion then use a string concat to add it. While not the most elegant solution this will suffice for now and we can rack this up as tech debt.