Adding End Shift Command
This command returns a Deferred object to manage the async nature of its operations, asynchronously getting the users location, combining this with the current time parsed into ISO8601 format before passing to the ShiftsRepository.

Because we support older versions of Android we have a problem formatting ISO8601 dates, in that there is no defined format that will work across OS versions. For older versions the closest we can get is to generate the date without the : in the timezone portion then use a string concat to add it. While not the most elegant solution this will suffice for now and we can rack this up as tech debt.

This command is almost identical to the start command. While they do share a lot of the code they will change for different reasons so I think it is reasonable to deal with the duplication. We could in the future write composed commands for formatting the date for example - especially if formatting to ISO8601 becomes a command problem. Another solution would be to implement it as an extension function and wrap some test around that. For now I am happy to deal with the duplication in order to crack on