Adding LocationRepository Interface
In order to start and end a shift we need to know the users location. This will have to be provided by the platform (Android). Because we want our domain layer to be platform agnostic we can simply define the interface in this module and allow other platform dependant modules to provide the implementation. We also define some common, expected exceptions that we can choose to handle or ignore in our clients

We define LatLng as type alias to make the API more readable (than Pair<Double,Double>). We could have chosen to implement this as a simple data class but because they are always returned as a pair this seemed appropriate. A class called LatLng is unlikely to contain any further data in the future. The only downside is no getLat/Lng property. However we still get kotlin proeprtyt destructuring