Adding ListenerUtils
This class provides 2 mapping utilities for when we deal with Events and Listeners. For Events we will often want to transform the value stored inside the event, while maintaining its error value. By writing a map extension method we can simply provide the mapping implementation as a higher order function and have the extension handle the boiler plate of null checks and Object creation.

For the function mapping, we will often need to deal with mapping of listeners. Our presentation layer will provide a listener with the type it wishes to receive, while our domain will need to subscribe to a repository that exposes entities and handle the mapping rules. This method provides a useful way to reduce boilerplate by simply mapping a mapper object to the original listener to receive a listener of the mapped type. The generic types make it slightly difficult to read, however when used inline it should provide a more fluent api