Adding Shift data object
These objects represent the data associated with a shift exposed from the domain to the presentation layer. in keeping with a clean architecture we wasn to avoid exposing our entities to our presentation layer, in order to simplify the presentation layer, encapsulate business rules, and decouple code in general.

A sealed class is a good choice to represent the state of a shift, because it allows us to provide an enumerable data object that itself can contain data. This means in the view if we are to switch on this type we will get a compiler error if we do not handle every available case. This can be useful when adding states at a later date. I chose to represent the start and end dates as Date objects as to hide the implementation of the service. We will still need to do some date parsing in the presentation layer but the way the dates are displayed is more of a view concern than a domain concern anyway. If a designer simply wishes to use a new format, only the presentation layer need change. if we are to use the domain across multiple applications, they can each display the date as the see fit.

I chose to expose the imageUrl as a URL object because this will ensure the string we receive is a valid url and can be loaded. Invalid URLs from the service will cause a parse exception. Our business rules can choose how to handle this. At the moment our only real choice is to return a default image url. Depending on the business needs we could make the URL nullable to declare to clients a lack of an image, or we can simply crash.