Adding ListenForShiftUpdatesCommand
This command provides an API for our presentation layer to listen for updates to our list of shifts. The API requires us to provide a listener for a list of Shift objects. We then register a listener with the repository and map the events to the output data class before passing them to the provided listener.

We provide an Ignore function as a return type to allow the clients to unregister. We could not have the call another command with the same listener to unregister, because the listener actually registered is the entity listener created inside the command.

Because we use weak listeners, not unregistering would not leak downstream objects indefinitely. However it is good practice to provide a way for clients to signal they are no longer interested in updates.

Because the listener signatures were getting quite verbose and hard to read I created a type alias Ignore (so called as it is the opposite of listen) for our ()->Unit functions. This again helps with the readability of the API.

Again, dealing with ISO8601 dates is a real pain in older versions of Android (yeah, thanks). I took the same approach as previously and manually manipulated the date String into a parsable format (removing the : in the timezone portion) before parsing it