Copying TestHelpers to presentation package
In order to use TestHelper functions int he presentation package we need to get the classes on the class path. The best way to do this is though gradle. You can export your test artifacts from one module and have other depend on them. The other way is the have a seperate module of test helpers you import for you tests only in any module that needs them. In the end both of these are a bit of work an d I simply do not have the time so I am copy pasting and racking this up as tech debt.