Adding ViewModelTestHelper
An  important part of testing ViewModels is testing their tear down, which happens in the onCleared method - handled by Androids lifecycle methods. However this method is package protected which means we cannot access it directly from our test. This extension circumvents this issue by exposing a public method in the same package and thus having access. This hack is a good example of why there is no package protected in kotlin, but that is a completely different discussion