Adding ShiftsViewModel
Exposes a very simple API that manages view state. We firstly expose a list of shifts that we update with events from our injected listenForShifts function. We then expose a start and end method to proxy calls to the injected start/endShift functions. These methods take a nullable view in order to aid binding to ClickListeners using method references in xml. We also manage state in regards to only executing one start/end action at a time, based on a private shiftUpdateProperty and the public shiftState property.

I create a CurrentShiftState enum to handle this (as we already have ShiftState, which is associated to a specific Shift object) This value can be used to control the display of controls (Button?) for starting / ending a shift.

We also expose 2 SingleLiveEvents for passing handling back to our client (fragment/activity) to handling things outside of the scope of this class. These are showing a generic error message dialog (when we get an unexpected error) and also requiring permissions.

Finally it is very important to call he returned Ignore function from listenForShifts. Failure to do this will potentially leak the ViewModel and associated context.