Adding AndroidLocationRepository
A LocationRepository implementation that uses the play services to get the users location. Although we execute the request as a Deferred coroutine on the default dispatcher we need to switch back to the Main thread in order to register our listener. We can the use the suspendCoroutine method to suspend the coroutine until we either receive a result, or error, either resume or resumeWithException Explicitly

Testing OS components like Locations can be quite difficult. I have excluded any unit tests at this point in order to focus on more pressing parts of the code. Strategies to test this include a full blown connectedTest or more ideally a simple Unit test. Whether we can unit test this will depend largely on whether we can mock the location provider. As far as checking permissions we could potentially inject a function that specifically handles this very Android specific part of the logic, or try to use some static mocking on ContextCompat