Adding CoarseLocation Dependency
We need this dependency in order to find the users location. I chose the COARSE location over the fine as this will allow us to use faster APIs when gaining the users location. If the business rules dictate that we need a much finer access we will need to update this and also most likely the implementation that uses it