Adding EventEmitter
In order to build a reactive API we need a way to add and remove listeners. We could implement this on a per client basis, but having an interface solves a few issues. Firstly it reduces the need for boiler plate code by providing generic implementations that can be composed. Secondly it allows clients to use the interface but delegate to other implementations that are injected. This gives us greater flexibility in our testing

We also need a data type to handle errors. For example, if we make a request that fails, listeners should be informed of the failure. This is what the Event<T> object aims to do, it contains 2 nullable properties, value and error, but provides constructors that dictate that at least one must be non null. Unfortunately this means we cannot implement it as data class, as the secondary constructors must call a primary constructor which would allow us to create an event will a null value and error. This is why we have had to implement an d autogenerated equals and hashcode.