Adding Id to shift data object
We will need a way to do a diff on each shift in the presentation layer - specifically for use with a diff util. The simplest way is to simply expose the shifts id. This way we know that 2 different shift objects actually represent the same shift but in a different state