Adding Dagger Dependency and Annotation processors
Dependency inversion is one of the most powerful tools in an engineers tool kit. It allows us to write testable, modular and configurable code. It is what allows us to write domain specific code that is not aware of the platform it is running on, define the details at a later date, and change out mind about those details whenever we like. We can also provide modules with different implementations when it suits us, for testing, for different environments etc

Dagger is the default dependency injection framework for Android - however it is not directly dependant on Android and is really a pure java framework. This means we can choose to include it in our domain without exposing it to our platform of choice