Adding Load Shifts Command
We need a way to expose this functionality to our clients without exposing any internal details or the entity layer. This command can be injected using its functional interface and provide a straight proxy to our repository. Allowing clients of the domain to complete this action without knowing about the implementation details