Handling 'null' response for no shifts
Because the API returns 'null' instead of [] for no shifts we cannot rely on using GSON to parse the response as it expects a valid JSON response. We can get around this issue when using retro fit by  using a custom ResponseConverter. However, we need a way to identify which type to apply it to, and because our previous response was List<ShiftEntity> we need a new type - this is because of type erasure at runtime our type will become simply List which is not enough detail. By making ShiftResponseEntity extend ArrayList, which implements list, we can keep all other code the same.

Notice that this code is included in the app package. This is because we (as much as possible) do not want our domain package to know about the implementation of the interfaces at its boundaries - it should not care that we are using retrofit (I made an exception with the annotations to reduce boiler plate). By providing this in the app / orchestration layer swapping out retrofit becomes a lot easier to manage

I also chose to move the ServiceFacory into the newly created networking package. Unlike the domain and presentation where we want to group classes by how they work together ie componentize them. In the App/Di/orchestration layer we are more likely to change things with what they do, because they are used more generically across usecases - ie networking, databases, location etc