Moving Location Repository to app component
Because the AndroidLocationRepository is a platform specific detail we need to provide and host it in a module with android dependencies (ie, not domain). We could host it in the presentation layer but this seems wrong for two reasons. Firstly it is not strictly a presentation component and secondly it is likely to be used across presentation modules. Also added null check on result, which can happen in the rare case the device has no last stored location (such as an emulator)