Adding Service Module
This is where we declare all our networking dependencies. Keeping all our networking dependencies in one module keeps it nicely encapsulated and easy to reason around. We could potentially simplify this as the only real external clients of this module are the service implementations, we could have 2 methods, one that provides the service factory (which will be used across the various services you will use) and then ShiftService. This would discourage people using GSON, OKHttp independently, but on the other hand provides a cleaner API