Adding ShiftsDomainModule
This module is used to provide all the dependencies for the Shifts component - which currently means the whole module. What is important to note here is that we provide all the "components" dependencies together. This makes it easy to reason about the dependencies and works to group things that change together, together.

We provide an ActivityScope Scope, which may go agains the idea that the domain should not know about where it will be used, but this scope could easily be called something else - I left it as activity by convention.

The start/end shift commands need to be provided with  qualifiers as they have the same interface.

We also need to use @JvmSuppressWildcards annotation. This is because of the way dagger works, reading the source files to generate new files, and the way it chooses to interpret kotlin generics, but create java code. It has a lot of issues with co/contra variance. one way is to wrote your providers in java, but even this can sometime lead to confusion an d require @JvmSupressWildCardAnnotations in the source. I think in order to keep component dependencies in one module, and write it in kotlin, it is best to simply use the annotation in the dagger modules. Its Ugly and difficult to manage at times, but at least it is contained to the Di, and does not affect source code. We keep the dagger problem within the scope of where we use dagger - the module