Adding ViewModelFactory
In order to use Dependency Injection with Android ViewModels, you need to provide a custom factory. It has a method that takes a class, and is constructed with a map of class to providers. This allows Android to create the ViewModels using your own providers, but allows the system to manage their lifecycle. We provide the ViewModelKey to be the key to the map of classes to providers