Adding Data binding
Data binding is a really powerful tool, especially when used with LiveData / ViewModels. It allows you to decouple the view model logic, from view component logic and presentation logic. You write VMs to handle the views model, custom views to handle presenting complex data, and your data binding adapters sit in the middle transforming and/or setting the models data onto the view. Because this layer is often simple get and set, with some basic transformation the data binding adapters are highly reusable, reducing large amounts of boilerplate code and freeing updactivities/fragments to deal with the android framework and lifecycle events