Adding CommonBindings
This class can be used to host extremely generic bindings. The initial implementation includes the include binding that binds a boolean to either View.VISBLE or View.GONE on a given view instance.

fadeVisible does the same, but first animates the view to 0 alpha (if false) before updating visibility. Because of the stateful nature of an animation we need somewhere to store them. This is what the animators map if for, but it is important to not that the references are Weak meaning we are not storing view references statically which would leak contexts and cause OOM exceptions