Adding Shift List presentation
We will show out list of shifts using a recycler view, as presumably the size of the list will grow fairly indefinitely. I chose to use 2 view holders and extending an abstract view holder. This makes the binding of the view holder from the adapter more simple as it simply needs to call the abstract method without caring about which subclass it is dealing with.

I provide ShiftsBindingFile that will be used to encapsulate all the ShiftsSpecific bindings - these will be bindings to concrete final types, and will not be reused outside of this component. The initial method we need is a way to display the start end date, so we use a simple date formatter if the date is present. We need to always have our data provided as nullable types as often binding can occur before MutableLiveData properties have been initialized. This should not happen in the adapter but it is good practice to follow this pattern to avoid NPEs

UI left extremely basic at this point as I want to focus on other more important aspects - obviously not production ready.