Updating ShiftsActivity to use data binding with injected view model
Rather than inject the VM directly we need to inject the factory and fetch it using the class as the key. This allows android to manage the lifecycle giving us back the same instance even if it is a different activity (through device rotation for example). We also need to make the call to AndroidInjection.inject, important to note that this is called before onCreate.

Another common gotcha is forgetting to set the lifecycle owner of the bindings. Forgetting to do this results in updates to the bounds data not being presented