Adding Shift Image Size
The images supplied from the API are square images, however we do not want to display them at full size (currently 500/500) so we should not use wrap content on the image views. Instead we can simply define a size as a dimen and use that. This also gives us scope to provide larger images for different screen sizes