Refactoring ViewModel to use suspending functions
Finally, after refactoring out domain layer we need to update our view model to take the new command Signatures - Action and ReceiveChannel. The principles remain largely the same, except we do not need to call await or handle our subscriptions, other than remembering to cancel the coroutine context. We do need to start using a TestCoroutineContext in our tests in order to handle the timing of the actions and make assertions at the correct time. I also had to move some of the state changes outside of async blocks to make assertions easier to handle without exposing the returned Job to our tests