Adding Action Interface
Because we want to move our domain towards using suspending functions over retuning Deferred objects but still want to inject our commands under and interface, we face an issue with dagger. because of the way suspending function work we cannot inject suspending lambdas (eg suspend ()->Unit). This is because as a java interface this is Function0<Unit> but when it is suspending actually compiles to Function1<Continuation<Unit>,Object>, under the hood kotlin adds an extra param which is the continuation. This means that dagger can never satisfy the provides method with what we expect to inject - unless we can some how control the order of compilation and annotation processing, which I'm not entirely sure is even possible.

To solve this we can simply use our own interface with the suspending keyword already added. As all our commands are currently ()->Unit, ie receive no value and return no value I decided to call the initial interface an Action as it seemed the most appropriate. As we need to pass and receive types we will need to create new interfaces. Unfortunately this will result in quite verbose type signatures, which is why I wanted to use the kotlin functions initially, but at the moment this seems to be the only way to get this working